rearranging, no functional change
